requirements.txt
streamlit==1.53.0
streamlit-folium==0.24.1
folium==0.19.4
ultralytics>=8.4.41
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
torch
torchvision
python-dotenv
